created game logic in main.js
created index.html and styles.css with no other changes made
Game Logic needs improvements. Need to handle exceptions better to quit
the program when user enters a keyword other than rock paper or scissor
I want the user to be able to quit the program as well with either q or
e.